requirements.txt
pylint==1.9.2
yamllint
black
PyYAML
pytest
